+++ apps/mobile/src/screens/HomeScreen.tsx
@@ -1,11 +1,60 @@
 import { Image as ExpoImage } from 'expo-image'
-import React, { useCallback } from 'react'
-import { Text, TouchableOpacity, View } from 'react-native'
+import { MotiView } from 'moti'
+import React, { useCallback, useReducer } from 'react'
+import {
+  Pressable,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View
+} from 'react-native'
+import { useNotifications } from 'react-native-notificated'
+
+import { StatusBar } from '../components'
+
+const styles = StyleSheet.create({
+  shape: {
+    justifyContent: 'center',
+    height: 250,
+    width: 250,
+    borderRadius: 25,
+    marginRight: 10,
+    backgroundColor: 'white'
+  },
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    flexDirection: 'row',
+    backgroundColor: '#9c1aff'
+  }
+})
 
 export const HomeScreen = (props: HomeScreenProps): JSX.Element => {
   const {
     navigation: { navigate }
   } = props
+  const [visible, toggle] = useReducer((s) => !s, true)
+  const { notify } = useNotifications()
+
+  function Shape() {
+    return (
+      <MotiView
+        from={{
+          opacity: 0,
+          scale: 0.5
+        }}
+        animate={{
+          opacity: 1,
+          scale: 1
+        }}
+        transition={{
+          type: 'timing'
+        }}
+        style={styles.shape}
+      />
+    )
+  }
 
   const navigateToDetails = useCallback(() => {
     navigate('Details', { id: 'home-id' })
@@ -17,10 +66,23 @@ export const HomeScreen = (props: HomeScreenProps): JSX.Element => {
         source="https://picsum.photos/seed/696/3000/2000"
         contentFit="contain"
       />
-      <Text>{'hello'}</Text>
-      <Text>{'thanks'}</Text>
-      <Text>{'app_information'}</Text>
-      <TouchableOpacity onPress={navigateToDetails}>
+      <Pressable
+        onPress={() => {
+          toggle()
+          notify('success', {
+            params: {
+              title: 'Hello',
+              description: 'Wow, that was easy',
+              hideCloseButton: true
+            }
+          })
+        }}
+        style={styles.container}
+      >
+        <StatusBar style="auto" />
+        {visible && <Shape />}
+      </Pressable>
+      <TouchableOpacity style={{ padding: 30 }} onPress={navigateToDetails}>
         <Text>{'home_screen.details'}</Text>
       </TouchableOpacity>
     </View>
